[IOS-959] Fixes Add Wallet screen showing BRD when already added
**Problem**

Despite adding BRD wallet to the home screen, BRD was still showing up
in the list of available wallets when tapping `Add Wallets`.

The problem was case sensitivity between the BRD wallet address in the
available wallets and in the previously added wallets. It appears that
BRD was the only such case.

**Solution**

When filtering out previously added wallets, lowercase the addresses being compared
(they are all ETH/ERC20 addresses so this is safe). Added a new function in `CurrencyListMetaData`
and associated unit test.

**To Test**

Load up a new wallet, tap `Add Wallets` and verify that the BRD token is not present
in the list of available wallets.

Go into Manage Wallets, hide the BRD token and verify that it doesn't show up on the home
screen, and that it still does not show up when you tap `Add Wallets`.