[IOS-946] Fixes crash on startup (saveBlocks())
Voisine hypothesized that surrounding buffers with `[...]` array literals
in calls to sql_bind_blob() was causing memory access errors in saveBlocks().

This was proven when Ed sync'd a wallet known to have a transaction with a hashes
array that was being called with the array literal wrapping the buffer.

The change is to remove the array literals and also pass the pointer address
for the non-array values such as `blockHash` and `flags`.

I performed three full BTC syncs with the same wallet, in both Debug and Release
mode in the simulator and in Debug mode on a device. All three syncs completed
without issue and restarting the app several times also succeeded.

Bumped version to 3.6.1 377