[IOS-993] Include current and previous console logs with Send Logs
**Background**

The developer menu option Send Logs was only capturing the console output from the current instantiation
of the app, because whenever the app started up it would reopen and overwrite the contents of log.txt.

This meant that if the app were to crash and we asked a user to send logs, we'd potentially miss clues
from the previous run of the app in which it crashed.

**Solution*

- on startup, check for content in log.txt and save this to previouslog.txt before overwriting log.txt
- when attaching the logs to the mail composer in `MessageUIPresenter`, attach both the current and previous logs

**To Test**

- run the app, hit Menu / Developer / Send Logs
- there should be one attachment for the current log
- shut down and restart the app, go back to Send Logs, and you should see two attachments
- change the To: field to send to your own email, and examine the attachments to verify
  that they have different content
- restart the app again and verify that you can send two new attachments; the content of the
  log marked *previous* should match what was in the log marked *current* (except possibly
  a bit at the end of the logs -- I believe this is due to some variability in when the console
  logging is actually flushed to log.txt)