[IOS-1070] Fixes 'x' not working for paper key screens
This was a regression due to the changes in [IOS-1070](https://breadwallet.atlassian.net/browse/IOS-1070)
in [this MR](https://gitlab.com/breadwallet/breadwallet-ios/merge_requests/136).

I added custom handling for the close button for the paper key screens (start and write) so that onboarding
events could be logged when 'x' was tapped. However, in other cases where the paper key screens are shown,
there was no longer any default dismiss handling.

**Testing**

- open the paper key screen from Menu / Security
- make sure 'x' works
- reopen it, tap "Write Paper Key"
- tap 'x' on the write-paper-key screen, make sure it dismisses
- run through the startup flow to the paper key step, make sure 'x' dismisses
  the screen