[NO-JIRA] Fixes rewards animation bug
Noticed this in my travels. Setting `shouldShowBRDRewardsAnimation` to `false` was only
being done after the shrink animation was complete, but the effect was that it would always
show the rewards animation if you did not linger in the BRD wallet until the shrink animation.

The fix is the set it to `false` after the expand animation so that we only ever show the expanded
rewards view once.

**To Test**

- reset user defaults
- re-run the app
- tap the brd wallet
- go back to home after the expand animation, before the shrink aimation
- re-enter the brd wallet, and make sure the rewards view is shown in its
  shrunken state