[IOS-1085] Decouples event caching from event sending
We have a long-standing issue that analytics events are not captured/uploaded
prior to having a PIN/wallet.

If the user bails during onboarding without entering a PIN, we still don't have a viable
solution. However, we can ensure that pre-PIN-entrty onboarding events are captured and
uploaded if the user does make it past the PIN entry.

To make this happen, we need to decouple the event gobbling and persisting from the
event uploading that happens once there's a wallet.