[IOS-1086] LES: ensure wallet has been created before updating transactions
On startup there can be a race condition where `updateTransactions` is called for a currency
before the associated wallet has been created.

This change guards on the existence of the wallet before proceeding with processing the transactions.

It does not necessarily (not likely) fix the issue cited in IOS-1086 but the potential race condition
was discovered while debugging general LES issues.