[IOS-1159] Fixes launch screen background color
The launch screen story board hard codes the bg color and this was not updated
when the new `primaryBackground` was introduced as part of the recovery key project.

Also updated the back button asset on the pin pad for the same reason.

To test, fire up the app and verify that the launch screen bg color matches the new, darker
background used on the login screen.