[IOS-1210] Fixes Touch ID/Face ID prompting
This was a regression from the Announcments refactoring, so that didPrompt() for
standard prompts was not setting the UserDefaults flag for biometrics.

**Testing**

Run the app on your device, with a fingerprint added, but Touch ID disabled within Menu/Security/Touch ID.

Get the Touch ID prompt to display, dismiss it, then leave and return to the home screen again and verify
that the prompt does not show up again. You can further verify by clearing UserDefaults via the developer
menu, then repeating the process above. Make sure the paper key prompt is suppressed either via the developer
menu or by writing down the key.