[IOS-1212] Biometrics spending limit screen text color fix
This screen will be going away with the work being done for IOS-564 (combining unlock and send
transactions for biometrics settings).

In the meantime, if a user opens this screen, they can't see the table view cell text due to lack
of contrast.

This is a quick fix to change the text color to white (primary).