[IOS-1224] Fixes fiat input for KRW, JPW, etc.
**Problem**

For currencies with high crypto/fiat ratios such as KRW and JPY,
floating point errors could cause fiat input such as "5" in the
send dialog to show output such as 4.99... or 5.02...

**Solution**

This is more of a fiat display formatting issue than a math issue; it turns out
that just setting the currency/number formatter's locale to match the locale associated
with the fiat rate solves the problem.

**Testing**

Test fiat input in the send dialog with various combinations of display currencies
and crypto currencies. In particular test KRW, KPW and JPY.