[IOS-1270] Use BRBrowserViewController with a URL for the ATM map
We have been pointing to the `/map` endpoint up until now, but that poses a security
risk on the platform side. Instead we'll now point a BRBrowserViewController directly
at the map URL.

Also includes some styling changes to the browser, removing the top and bottom bars
and setting the status bar to light since the map background is dark.

Finally, the map experiment now includes a meta parameter which is now attached
to the Experiment object so that the map URL can be specified (and potentially changed)
as part of the experiment.