[IOS-1169] Wires up wallet sync progress
The new sync progress interface is a beautiful thing.

This MR grabs the sync progress and timestamp from Core and passes them
up to the UI layer, eventually landing at the SyncingIndicator.

To test, run the simulator with one or more unsync'd wallets and make sure
the home screen cell shows a percent. Enter the wallet and make sure the
sync'ing header shows the sync'd-through date as well as the percent.