Fixes threading issue in support center web server.
I found this while running the thread sanitizer. We were calling `loadWebView()` twice on the
support center to preload the content. The second load required the existing HTTP server to be
shut down, which was happening on a different thread than the original init, according to the
thread sanitizer (it's a bit murky, because ApplicationController dispatches some asset bundle
processing to a background thread, but then throws the support center preloading onto the main thread
async).

Solution: Don't call `loadWebView()` twice; no need to call it from `init` and from the modal presenter.
Also renamed the preloading func to make things read a bit more intuitively.