[IOS-1294] Separated UI alert code from ModalPresenter
**Back story**

The modal presenter does a lot of things, most of which require the system to be fully
initialized, with key store, wallet, etc. Due to changes in the startup sequence related
to 4.x, the modal presenter needed to be created later in the process, and one side effect
was that it was not available yet when the PIN was set for a new wallet at the end of onboarding.

The modal presenter also does things like generic UI alerts that are not dependent on the system
being fully initialized.

This MR pulls out the alerting functions into a new class called `AlertPresenter`, including the
reachability alert.

**To Test**

- create a new wallet, set the PIN, and verify that the PIN-set slide-up alert is displayed
- change the PIN for an existing wallet, and verify that the alert is displayed
- start alert in airplane mode and verify that the no-network alert is displayed and is cleared
  when the network is restored