Delays the push notifications prompt a little
We were showing the push notifications opt-in prompt when the home screen first appeared,
and there are a lot of other things going on at the time, such as the paper key prompt,
and the BRD wallet animation.

This MR adds a check to the should-show-prompt check in `NotificationAuthorizer`, requiring
that the app has launched at least twice before we show the prompt. This will make the post-create/post-restore
experience less chaotic.