Cleans up ModalPresenter on wipe
**Problemo**

When wiping a wallet, we were leaving strong references to the ModalPresenter in a bunch of places,
particularly menu item callbacks. This meant that ModalPresenter.deinit would not be called when the system
was reset and a new ModalPresenter was created. In turn, we'd have an old ModalPresenter still subscribed
to Store triggers, causing assertion failures.

**Solution**

- Fixed the closures that lacked weak references to the ModalPresenter.
- cleaned up the menu in the didWipeWallet handler

**To Test**

- put a breakpoint in ModalPresenter.deinit
- restore/create/wipe multiple times in the same simulator session
- make sure the breakpoint is hit when you create/restore for the second time in a session
- I ran into a Core assert on my third consecutive wipe, on a freshly created wallet, as it was
  still doing sync'ing of ETH transactions, so you might run into a similar issue. Ed and Michael
  are aware of this.