[IOS-1374] Hides FAQ button on set-pin screen during onboarding
During onnboarding we don't yet have a wallet, therefore we can't authenticate
platform requests. The support center FAQ pages use `BRWebViewController`, which requires
an authenticator.

Adddressing this and decoupling authentication from the browser view controller would require
a significant refactor, so to solve this, we're opting to just hide the FAQ button on the onboarding
set-pin page. Its utility is debatable anyway.

To test, run through onboarding and verify that there is no FAQ button on the set-pin screens.