[IOS-1387] Fixes Go to Wallet issue after creating recovery key
If you write down the recovery key at the end of onboarding, the Congratulations screen says
Go to Wallet. This makes sense. However, if you write down the key later from the menu, Go to Wallet
doesn't go back to the home screen; instead it just dismisses the recovery key flow and leaves you at the menu.

This MR hides the Go to Wallet button if the write-key flow is not accessed via onboarding, and includes a close
button ('x') instead.

To test, write/confirm the recovery key in the following scenarios:

- at the end of the onboarding
- from the home screen prompt
- from the security menu

Verify that Go to Wallet is only shown in the first case and that the close button works in the other cases.