[IOS-1391] Fixes crash during wipe
Dee encountered one crash while testing wallet wipes, and I reproduced it where
we were running a block on ModalPresenter after it was dealloc'd.

This fixes this order (once and for all?) of diconnecting, menu cleanup, etc. when
wiping a wallet.

To test, restore then wipe a wallet via the developer menu or security settings.
Make sure the wallet is wiped, the home screen does not appear briefly before
the startup screen is displayed.