Change to lazy var for sync percent number formatter.
updateTextLabel() on the sync indicators gets called hundreds of times, and we
were allocating a NumberFormatter each time. There was a crash in [NumberFormatter init]
in iOS system land, reported by Kevin Ponto on TestFlight. That shouldn't happen, but avoiding
so many allocations may mitigate the problem.

While I was at it, made the asset total number formatter on the home screen a lazy var as well.