[IOS-1160] Fixes Japanese recovery key confirmation issue
When entering the last character of the second recovery key confirmation word,
the keyboard was dismissing, as it should, to expose the Continue button. However,
when inputting Japanese characters, the last character was being trimmed (not sure why,
but the system was calling textChanged() with the last character missing).

I experimented a little and found that not auto-dismissing the keyboard prevented
this from happening. Taking that idea further, I queued the keyboard dismissal with
a DispatchQueue.async and that did stopped the issue from happening but preserved the
auto-dismissal behaviour that we want.