[IOS-1438] Fixes About text overflow
The about text "Made by the global BRD team..." was outside the left/right margins
for certain languages such as Spanish.

Fix:

- made it a wrapping label
- added left/right margins

To test, hack the footer text at `AboutViewController.swift:87` to use a really long string,
then view the About screen and verify that the text wraps.

Alernatively, refer to the screenshot.