IOS-1507 Showing send modal twice
Reason for this bug was that `ModalPresenter` is retain cycled, hence it is never unsubscribed. When wallet is unlinked new instance of `ModalPresenter` is created. Resulting in two instance being subscribed and presenting modals.  I have tried to fix the leak itself, however profiler shows 288 leaks. This makes finding the leak non-trivial. We should create a ticket to deal with the leaks. For now Im manually unsubscribing the instance that should be deinitialised.